+++ src/backend/config/ConfigOptions.h
@@ -67,6 +67,7 @@ public:
 	static QString rotateWatermarkEnabledString();
 	static QString stickerPathsString();
 	static QString useDefaultStickerString();
+	static QString showControlsWidgetEnabledString();
 	static QString captureCursorString();
 	static QString snippingAreaRulersEnabledString();
 	static QString snippingAreaPositionAndSizeInfoEnabledString();