+++ src/backend/config/IConfig.h
@@ -157,6 +157,9 @@ public:
 	virtual QColor canvasColor() const = 0;
 	virtual void setCanvasColor(const QColor &color) = 0;
 
+	virtual bool showControlsWidget() const = 0;
+	virtual void setShowControlsWidget(bool enabled) = 0;
+
 	// Image Grabber
 
 	virtual bool isFreezeImageWhileSnippingEnabledReadOnly() const = 0;