+++ src/gui/MainWindow.cpp
@@ -630,6 +630,9 @@ void MainWindow::setupImageAnnotator()
 	mImageAnnotator->setNumberToolSeedChangeUpdatesAllItems(mConfig->numberToolSeedChangeUpdatesAllItems());
 	mImageAnnotator->setStickers(mConfig->stickerPaths(), mConfig->useDefaultSticker());
 	mImageAnnotator->setCanvasColor(mConfig->canvasColor());
+	if(mConfig->showControlsWidget()) {
+		mImageAnnotator->showControlsWidget();
+	}
 }
 
 void MainWindow::captureDelayChanged(int delay)