+++ src/gui/imageAnnotator/KImageAnnotatorAdapter.cpp
@@ -179,6 +179,11 @@ void KImageAnnotatorAdapter::showRotator()
 	mKImageAnnotator->showRotator();
 }
 
+void KImageAnnotatorAdapter::showControlsWidget()
+{
+	mKImageAnnotator->showControlsWidget();
+}
+
 void KImageAnnotatorAdapter::setCanvasColor(const QColor &color)
 {
 	mKImageAnnotator->setCanvasColor(color);